refactor: streamline build process for precompiles
